+++ layers/Engine/packages/graphql-parser/tests/ExtendedSpec/Execution/MultipleQueryExecution/MultipleQueryExecutionEnabledTest.php
@@ -2,7 +2,7 @@
 
 declare(strict_types=1);
 
-namespace PoP\GraphQLParser\ExtendedSpec\Execution;
+namespace PoP\GraphQLParser\ExtendedSpec\Execution\MultipleQueryExecution;
 
 class MultipleQueryExecutionEnabledTest extends AbstractMultipleQueryExecutionTest
 {
